Generate form-data.h from yaml
This converts the data defined in form-data.h into a series of yaml files
in dat/forms (similar to what was done for monsters).

The data blocks in form-data.h were very opaque to read or edit, and this
should be a lot clearer (while also making future structural changes
easier, since default values for new fields can be automatically propogated
to any existing form, with less risk of typos).