Fix Paragon sometimes losing its link to the player (salticid)
The duration of the paragon link status (which was used to determine
whether you could have attacks against you blocked by the paragon, as well
letting your attacks trigger the paragon and recasts perform its super
attack) was set to the same duration as the paragon's own timer, with the
intent that it last exactly as long as the paragon did.

However, it was possible for the paragon to end up lasting longer (more
easily if it was slowed), resulting in a period of time where the paragon
was visibly active, but you gained none of these benefits from it (and
recasting the spell would just replace it, rather than allowing you to use
its super move).

Now it doesn't attempt to expire at all, but simply end when the paragon
dies (which it always *also* did - meaning the original duration may have
always been unnecessary).