Fix shapeshifter monsters spawning in their natural form (Ge0FF)
Broken somewhat subtly by 31604a8. monster_polymorph will fail if the
monster isn't in a valid position (since the space they're on isn't
habitable for anything they could turn into), so moving when in
_place_monster_aux the monster is actually put on the grid caused them to
always spawn in their native, unshifted form.

I have moved the polymorph lower in the function. I am a little concerned
that there may be other non-obvious consequences to the change done in
31604a8, but a cursory inspection didn't turn up anything else that seemed
dependent on the monster having a real position.