Change/improve the talisman form properties display
It now displays many scaling values in a table, showing their value at
min, max, and current shapeshifting skill (which should hopefully give
players a better idea of how much each scales).

It also lists resists, which slots are melded, and which categories of
mutations are suppressed. (Along with a handful of form-specific things
like blinkbolt damage and vampire healing chance.)

And, in general, the information is presented in a more compressed format
that is easier to read at a glance.

Talisman database entries have most lines about their slot-melding removed,
since they are now redundant with the more explicit listing right below
that.

Many Form methods are refactored to accept a specific skill value (with -1
defaulting to 'the player's current skill')