Replace Beast form with Quill form
Beast form was, more or less, a glorified ring of slaying in exchange for
your aux armour. And while this could be a useful (if bland) trade in the
earlygame, it felt to me like a bit of a waste of the traditionally very
popular 'werewolf' fantasy. So I am planning to promote werewolf into a
higher-tier form where it can be a bit more exciting, and replace beast
with something else of suitable power level for shapeshifters to start in.

Quill form also trades all of your aux armour away, but in return it makes
you spiny (with damage that scales with shapeshifting). 40% of melee
attacks made against the player will trigger an automatic retaliation for
1d2 - 1d8 damage (depending on shapeshifting skill). This is sometimes
stronger than beast form's slaying and sometimes weaker, but also more
distinctive (and it makes the shapeshifter start more distinct *visually*
too, in my opinion - selling the idea that you're a weird metamorph instead
of just 'a slightly different colour adventurer').

-----------------------

The quill talisman and form tiles are by regret-index.
 * The quill talisman combines a CC0 morningstar by idylwilds with an
   unused tile for the Sceptre of Asmodeus by Denzi.
 * The humanoid form combines Sastreii's beast form tiles with roctavian's
   old porcupine tile. Djinn combine this with Team Impy's djinn player
   doll tile, Armataurs combine this with Sastreii's armataur player doll
   tile, and Naga combine this with Sastreii's naga player doll tile.
 * Quill form felids use Denzi's old cat tiles with roctavian's porcupine.
 * Quill form octopodes combine roctavian's porcupine tile with Porkshop's
   edit of dolphin's old octopode tile.