New form: Living Hive
This is a statue-tier form, intended to diversify appealing options at that
tier. It also melds body armour (and helm), providing a much smaller amount
of AC, but also a significant amount of both HP and MP regeneration (which
are improved by shapeshifting skill). In addition, whenever the player's HP
falls below 50%, multiple angry bees will swarm out from your body to
defend you. These come naturally berserk (and with the Concentrate Venom
status!) to make something like a killer bee still relevant into lategame.
Like Mark of the Celebrant, this status recharges only once the player
reaches full HP again.

The hope was to provide something with enough defensive utility to be
attractive to some of the characters who would otherwise consider statue
form their main viable option, while being as different as possible. The
MP regeneration it provides (and not slowing the player!) also makes it
markedly stronger as a caster form than statue is (and potentially of
interest to characters who'd not otherwise consider training shapeshifting
at all!)

And who doesn't want to unleash swarms of bees upon their enemies?

(Okawaru disapproves of this form - not only because the swarm of allies is
a core part of it, but the regeneration is even thematically coming from
the labor of your little friends and we all know that Oka hates friends.)

-------------------

The hive form tiles and preliminary honeycomb talisman tiles were made by
regret-index.
 * The initial talisman combines Denzi's old honeycomb and amulet tiles.
 * The hive form tiles use CC0 honeycomb tiles by delta12 studio. The
   felid tile mixes it with Denzi's felid tiles and the octopode tile
   mixes it with Bloax's base octopode player doll tile.