Let Fighter start with a second potion of might (Lici)
Of all the backgrounds, it's arguable that fighters start with the smallest
kit in total. Most everyone else starts with *multiple* consumables (or a
book on top of their consumables). Having more stuff is more fun (and
further distinguishes melee starts from each other), so let's give them an
extra.