Fix some wierdness with poltergeists and Fisticloak
The most obvious manifestation of this bug was the cloak sometimes falling
off when the player took the stairs, but the underlying issue is that if
you had 5 aux slots filled up already, and tried to put on the fisticloak,
it would still let you *even though it should require two of those slots*.

The issue is that the search for both 'a slot that can hold a cloak' and
'a slot that can hold a hat' was finding the same single slot and thinking
this was good enough. (And then save compat code was knocking off the
fisticloak on level change, since it properly detected that you didn't have
enough slots for it.)

Now, multi-slot items should substract each slot they take up from your
total slots, when calculating whether something can fit or not.

This fixes #4322