Allow overwriting a cloud with another cloud of the same type
(So long as the new cloud has a longer duration.)

This was aimed foremost at Rust Breath (which has a very long duration and
high cost, so being unable to place clouds on top of other almost-expired
clouds felt awkward.)

It does also affect a few other player spells, such as Fire Storm and
Mephitic Cloud. The latter is likely the largest effect of anything
player-facing, but it seems unlikely to me that this buff will have an
excessive effect on its power.

It does *not* affect Freezing Cloud (whose targeting causes it to spread
around existing clouds, rather than place in a fixed area.)

Likewise, it doesn't affect most monster cloud uses, which obey similar
rules (but does affect a couple, like catoblepas breath. This seems like it
should be fine to me.)