Revamp dragon talisman somewhat
Dragon form was always relatively unpopular due to its (somewhat
unintuitive) melee glass cannon nature. But beyond that, "You do a ton of
melee damage, but your breath weapon is barely even worth using." is a
terrible flavor fit for monsters who are most iconically defined *by* their
breath, in my opinion.

The breath weapon (as a remnant of being so ancient that it once was a
general mutation) also scaled with player XL and not shapeshifting skill in
any fashion, contrary to... basically every scaling form property these
days.

This commit aims to improve the situation in the following ways:
 -The default dragon type is now a golden dragon, rather than a fire
  dragon (per regret-index's suggestion). This arguably gives it a bit of
  a niche as a 'resistances' form to contrast with its overall mediocre raw
  defenses (due to its size).
 -The breath weapon is now *substantially* more powerful, but shares the
  same XP recharge system that draconian breath does (which seems to have
  been pretty successful.) It effectively fires all 3 golden dragon breath
  types at once, doing a piercing line of fire and cold damage, and leaving
  poisonous clouds behind. It scales with shapeshifting skill, not XL.
 -The UC boost has been meaningfully reduced (and default tail slap aux
  removed - though draconians retain theirs), with the idea that the
  breath should account for a meaningful chunk of its overall power.
 -Draconians get a substantial shapeshifting skill-based boost to their own
  breath power (instead of the form flatly doubling it), but it also
  recharges twice as quickly (to match the recharge rate of non-draconians
  using this talisman).

Numbers are a bit provisional on some of this, but the hope is that it
feels appealing to more people, and also more 'like a dragon' to play as.

-------------------------------

The golden dragon talisman, form, ability icon, and vfx were edited and
made by regret-index.
 * The talisman recolours the previous dragon-coil talisman using the
   older fire dragon design made by Denzi and edited by roctavian, Marbit,
   and Bloax.
 * The form tiles all recolour Sastreii's regular, felid, and octopode
   dragon form tiles.
 * The golden breath icon combines ontoclasm's Spit Poison, Breathe Fire,
   and Breathe Frost tiles.
 * The golden breath vfx combines Sastreii's light bolt with roctavian's
   Throw Fire and Throw Frost effects.