fix: an issue with subprocess and signal modules in Python 3.12
Launching the server gives the following error on some Python 3.12s:
"AttributeError: module 'subprocess' has no attribute 'signal'."

Previously, the subprocess module imported `signal`, so it was possible
to use `subprocess.signal`. But this `import` was removed from the
module in [1] in January. This caused some bugs, and the commit was
reverted [2] on Feb 16, half a month later. Some distributions, for
example Arch Linux, provide Python 3.12 with the commit [1], but
without [2] (yet), hence the error.

To fix this, import the signal module directly instead of relying on
`subprocess` to provide it, as was suggested in #4353. Also, the
subprocess module is not used for anything else in `process_handler.py`,
so it should be safe to remove.

This addresses the second part of #4353 and allows to run the server on
Python 3.12. Other supported versions of Python shouldn't be affected by
this change at all. Supporting 3.13 requires replacing the deprecated
crypt module, which should be done by someone more familiar with the
authentication code.

[1] https://github.com/python/cpython/commit/f65aa0d1
[2] https://github.com/python/cpython/commit/042b4924