Add Protean talismans (for Shapeshifter start)
Having a diversity of early/midgame forms is good for the variety of anyone
who might pivot into shapeshifting by finding these talismans during the
game, and it feels sad to canonize just one of them by putting it into the
shapeshifter start (and thus heavily limiting who is likely to actually use
it), but we'd hardy want to make shapeshifter start with 4 different
talismans of the same tier, either.

This is an attempt to randomize what capstone form shapeshifter will end up
with, while limiting the startscumming incentives that would come with just
giving the player a random talisman at the start. Instead, they are given a
protean talisman that *turns into* a random tier 2 talisman when evoked
with at least 6 shapeshifting skill (ie: slightly below the min skill to
use whatever it turns into).

(I have picked a mix of full-body and partial-meld forms for this tier,
since I've seen it reasonably argued that full-body forms make species much
less important to the background than it would otherwise be, but at the
same time, I feel like they're an important part of the shapeshifting
fantasy - to become something monstrous and very unlike one's normal form.
So this allows us to split the difference.)

Protean talismans are not intended to generate as floor items (though some
vaults could still be allowed to place them.)

----------------------

The protean talisman tile was made by regret-index, and combines
Sastreii's protean progenitor tile with ontoclasm's elemental wellspring
tile.