Mark maxed out misc evokers as useless if not in player's inventory (#4402)
Existing code only marks useless misc evokers as such if they're in
shops, but if they're just laying on the floor they are currently not
considered useless. This is a nuisance in zigs, which have one or more
evokers on almost every floor, quickly maxing them out and adding more
visual noise to an already overloaded ctrl-X menu.

The hacky inventory check is not ideal, but I couldn't come up with a
better solution without going out of my depth and messing with core
functions. I checked all edge cases I could think of like standing on +5
evo while having a copy in your inventory