Fix closing doors with the mouse (#4276)
Control left mouse clicking an open door would result in an unknown
command message instead of attacking in the direction of the door. The
reason for this is that the game would try to close an open door if you
control left clicked it and couldn't see a monster on it, but generate
the command to do so incorrectly. This is a bad idea as control left
clicking is almost always used to attack invisible monsters. This change
moves closing doors to shift left clicking and fixes the generation of
the command.

Fixes #3801