Quickly fit in some more 4/1 tiles (CanofWorms)
One of these was intended for magical staves, but the code matching tiles
for those as identifiable and visually consistent between types is a fair
bit of a headache to slip in temporary tiles for, so I instead gave it to
Yiuf.

While I'm at it, I also re-enabled a much older hat tile, since it suits
the time at hand.