Fix some issues with coglins wielding and unwielding weapons (#3880)
- Fix distortion unwield banishment triggering a turn late for coglins
- Fix delay interrupts (most notably abyss entry from coglin distortion
unwield) from asking the player twice whether they would like to "Keep
equipping themselves?"

Another thing that arose during the making of this PR: when being
banished or shafted, `you.time_taken` is set to 3/4 of the usual time it
would take for the player to complete their action, which can have
bizarre effects. Attacks and spells can be cast, items can be equipped,
scrolls/potions/wands can be used, all randomly and inexplicably 2-3
auts faster than usual. This feels a bit counterintuitive, and I
wouldn't be surprised if it has unintended effects.

There was a bug I experienced a while ago with ?revelation carrying over
to the new floor when being shafted the turn after reading the scroll
which I couldn't reproduce, but I wonder if this is related.